Changed Authority to implement Role based Authority

+++ src/main/java/com/beamsoftsolution/taxfm/repository/AuthorityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
+public interface AuthorityRepository extends JpaRepository<Authority, Long> {
 	
 	Optional<Authority> findByAuthority(String authority);
 }
